+++ test/util-file.test.js
@@ -6,10 +6,11 @@ const {
   decodeLargeNumbers,
   encodeLargeNumbers,
   getRemoteFile,
-  getLocalFile, stringify, addQuotesToRefInString, parseString, isJSON, isYaml, detectFormat
+  getLocalFile, stringify, addQuotesToRefInString, parseString, isJSON, isYaml, detectFormat, analyzeOpenApi
 } = require("../utils/file");
 const yaml = require('@stoplight/yaml');
 const {describe} = require("@jest/globals");
+const mockOpenApi = require("./__utils__/mockOpenApi.json");
 
 describe('openapi-format CLI file tests', () => {
 
@@ -409,5 +410,39 @@ describe('openapi-format CLI file tests', () => {
     })
 
   })
+
+  describe('analyzeOpenApi function', () => {
+    test('should extract OpenAPI info', () => {
+      const mockOpenApi = require('./__utils__/mockOpenApi.json');
+      const result = analyzeOpenApi(mockOpenApi);
+
+      expect(result).toEqual({
+        flags: ["x-customTag"],
+        tags: ["pets"],
+        operationIds: ["getPets", "updatePet"],
+        paths: ["/pets"],
+        methods: ["GET", "PUT"],
+        operations: ["GET::/pets", "PUT::/pets"],
+        responseContent: ["application/json", "application/xml"],
+        flagValues: ["x-version: 1.0"]
+      });
+    });
+
+    test('should handle null when extracting OpenAPI info', () => {
+      const mockOpenApi = require('./__utils__/mockOpenApi.json');
+      const result = analyzeOpenApi(null);
+
+      expect(result).toEqual({
+        "flagValues": [],
+        "flags": [],
+        "methods": [],
+        "operationIds": [],
+        "operations": [],
+        "paths": [],
+        "responseContent": [],
+        "tags": []
+      });
+    });
+  });
 });